​​refactor: asyncErrorBoundary 함수 fallbackComponent를 주입받도록 수정
