+++ src/components/ErrorBox/ErrorBox.ts
@@ -5,16 +5,6 @@ export const addErrorBox = (text: string) => {
   $movieListSection.replaceChildren($ErrorBox({ text }));
 };
 
-export const asyncErrorBoundary = async (asyncFn: () => Promise<void>) => {
-  try {
-    await asyncFn();
-  } catch (error) {
-    if (error instanceof Error) {
-      addErrorBox(error.message);
-    }
-  }
-};
-
 interface ErrorBoxProps {
   text: string;
 }
